requirements.txt
asgiref==3.9.2
Django==5.2.7
django-debug-toolbar==6.0.0
django-filter==25.1
djangorestframework==3.16.1
factory_boy==3.3.3
Faker==37.8.0
sqlparse==0.5.3
tzdata==2025.2
